pyproject.toml
[project]
name = "excelrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "flagembedding>=1.3.4",
    "google-genai>=1.11.0",
    "huggingface-hub[hf-xet]>=0.30.2",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "marker-pdf>=1.6.2",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "torch>=2.6.0",
]
